override material UI theme at root instead of on page
also changed prop name from array to options
